# Conflicts:
#	app/Http/Controllers/Pages/DashboardController.php
#	app/Http/Controllers/Pages/PostController.php
#	app/Http/Controllers/Pages/TutorController.php
#	app/User.php
#	resources/views/baidang.blade.php
#	resources/views/dangbaitimgiasu.blade.php
#	resources/views/dangkygiasu.blade.php
#	resources/views/giasu.blade.php
#	resources/views/page/header.blade.php
#	resources/views/page/slider.blade.php
#	routes/web.php

+++ resources/views/post/baidang.blade.php
@@ -45,12 +45,14 @@
 								<p><strong>Địa chỉ: </strong> {{$post->address? $post->address : '-'}}</p>
 								<p><strong>Yêu cầu: </strong> {{$post->requirement? $post->requirement : '-'}}</p>
 								
-								<button class="btn-post"><a href="post/{{$post->id}}/edit" class="link-edit"> Chỉnh sửa</a></button>
+								@if(Auth::user()->roleAdmin())
+									<button class="btn-post"><a href="post/{{$post->id}}/edit" class="link-edit"> Chỉnh sửa</a></button>
 
-								{{ Form::model($post, ['url' => ['/post', $post["id"]], 'method'=> 'DELETE', 'enctype' => 'multipart/form-data', 'style' => 'display: inline-block;',
-								 'onsubmit' => "return confirm('Bạn có muốn xóa suất dạy này không?');"])}}
-									<button type="submit"  class="btn-post">Xóa</button>
-								{{ Form::close()}}
+									{{ Form::model($post, ['url' => ['/post', $post["id"]], 'method'=> 'DELETE', 'enctype' => 'multipart/form-data', 'style' => 'display: inline-block;',
+									 'onsubmit' => "return confirm('Bạn có muốn xóa suất dạy này không?');"])}}
+										<button type="submit"  class="btn-post">Xóa</button>
+									{{ Form::close()}}
+								@endif
 								<hr>
 							</li>
 						@endforeach
